+++ src/index.ts
@@ -2,3 +2,12 @@ export {
   formatJson,
   type FormatJsonOptions
 } from "./formatter/format";
+
+export {
+  isValidJson,
+  JsonParseError,
+  parseJson,
+  validateJson,
+  type JsonValidationError,
+  type JsonValidationResult
+} from "./validator/validate";